refactor : Upload View Controller 상속으로 변경
